+++ cppdep.py
@@ -338,8 +338,8 @@ class Package(object):
         components: The list of unique components in this package.
     """
 
-    _RE_SRC = re.compile(r'(?i).*\.((?P<h>h(h|xx|\+\+|pp)?)|'
-                         r'(?P<c>c(c|xx|\+\+|pp)?))$')
+    _RE_SRC = re.compile(r'(?i)\w+((?P<h>(\.h(h|xx|\+\+|pp)?)?)|'
+                         r'(?P<c>\.c(c|xx|\+\+|pp)?))$')
 
     def __init__(self, name, group, src_paths, include_paths, alias_paths):
         """Constructs an empty package.